Bug 1978601 - Moves `StorageMaintenanceWorker` and `StorageExtensions` to `concept-storage`
