+++ mobile/android/android-components/components/browser/storage-sync/src/main/java/mozilla/components/browser/storage/sync/PlacesHistoryStorageWorker.kt
@@ -6,6 +6,7 @@ package mozilla.components.browser.storage.sync
 
 import android.content.Context
 import androidx.work.WorkerParameters
+import mozilla.components.concept.storage.StorageMaintenanceWorker
 import mozilla.components.support.base.log.logger.Logger
 
 /**
@@ -14,7 +15,7 @@ import mozilla.components.support.base.log.logger.Logger
  * If there is a failure or the worker constraints are no longer met during execution,
  * active write operations on [PlacesStorage] are cancelled.
  *
- * See also [StorageMaintenanceWorker].
+ * See also [mozilla.components.concept.storage.StorageMaintenanceWorker].
  */
 internal class PlacesHistoryStorageWorker(context: Context, params: WorkerParameters) :
     StorageMaintenanceWorker(context, params) {
