+++ mobile/android/android-components/components/concept/storage/src/main/java/mozilla/components/concept/storage/StorageMaintenanceWorker.kt
@@ -1,8 +1,4 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-package mozilla.components.browser.storage.sync
+package mozilla.components.concept.storage
 
 import android.content.Context
 import androidx.work.CoroutineWorker
